a simple pitch detector plugin
This build met my first goal for the plugin an ability to detect pitch. It is still taking up a bit too much CPU. The plan is to implment settings in the GUI to adjust the polling rate to align with midi and allow the PPQ rate to be set.   The next goal of the plugin is to include logging of the pitches to export to midi.   Then try to make a polyphonic version.